Use the correct route /profile
